save old stuff
moving to branch updated

+++ ivp/ivp_utility/ivu_set.hxx
@@ -38,7 +38,7 @@ class IVP_U_Set: public IVP_VHash {
   //friend class IVP_U_Set_Enumerator;
 protected:
     IVP_BOOL compare(void *elem0, void *elem1) const { return (elem0 == elem1)?IVP_TRUE:IVP_FALSE; }
-    int      elem_to_index(T *elem){ return fast_hash_index((long) elem); }
+    int      elem_to_index(T *elem){ return fast_hash_index((size_t) elem); }
 public:
   void add_element(T *elem){
       IVP_VHash::add_elem(elem, elem_to_index(elem));
